+++ aleph_alpha_client/explanation.py
@@ -12,6 +12,8 @@ from typing import (
 # Import Literal with Python 3.7 fallback
 from typing_extensions import Literal
 
+from aleph_alpha_client import Text
+
 from aleph_alpha_client.prompt import ControlTokenOverlap, Image, Prompt, PromptItem
 
 
@@ -204,6 +206,20 @@ class TextScore(NamedTuple):
             score=score["score"],
         )
 
+class TextScoreWithRaw(NamedTuple):
+    start: int
+    length: int
+    score: float
+    text: str
+
+    @staticmethod
+    def from_text_score(score: TextScore, prompt: Text) -> "TextScoreWithRaw":
+        return TextScoreWithRaw(
+            start=score.start,
+            length=score.length,
+            score=score.score,
+            text=prompt.text[score.start:score.start + score.length],
+        )
 
 class ImageScore(NamedTuple):
     left: float
@@ -236,6 +252,20 @@ class TargetScore(NamedTuple):
             score=score["score"],
         )
 
+class TargetScoreWithRaw(NamedTuple):
+    start: int
+    length: int
+    score: float
+    text: str
+
+    @staticmethod
+    def from_target_score(score: TargetScore, target: str) -> "TargetScoreWithRaw":
+        return TargetScoreWithRaw(
+            start=score.start,
+            length=score.length,
+            score=score.score,
+            text=target[score.start:score.start + score.length],
+        )
 
 class TokenScore(NamedTuple):
     score: float
@@ -275,23 +305,37 @@ class ImagePromptItemExplanation(NamedTuple):
 
 
 class TextPromptItemExplanation(NamedTuple):
-    scores: List[TextScore]
+    scores: List[Union[TextScore, TextScoreWithRaw]]
 
     @staticmethod
     def from_json(item: Dict[str, Any]) -> "TextPromptItemExplanation":
         return TextPromptItemExplanation(
             scores=[TextScore.from_json(score) for score in item["scores"]]
         )
+    
+    def with_text(self, prompt: Text) -> "TextPromptItemExplanation":
+        return TextPromptItemExplanation(
+            scores=[TextScoreWithRaw.from_text_score(score, prompt) if isinstance(score, TextScore) else score for score in self.scores]
+        )
+
 
 
 class TargetPromptItemExplanation(NamedTuple):
-    scores: List[TargetScore]
+    scores: List[Union[TargetScore, TargetScoreWithRaw]]
 
     @staticmethod
     def from_json(item: Dict[str, Any]) -> "TargetPromptItemExplanation":
         return TargetPromptItemExplanation(
             scores=[TargetScore.from_json(score) for score in item["scores"]]
         )
+    
+    def with_text(self, prompt: str) -> "TargetPromptItemExplanation":
+        return TargetPromptItemExplanation(
+            scores=[TargetScoreWithRaw.from_target_score(score, prompt) if isinstance(score, TargetScore) else score for score in self.scores]
+        )
+    
+
+
 
 
 class TokenPromptItemExplanation(NamedTuple):
@@ -352,6 +396,31 @@ class Explanation(NamedTuple):
             ],
         )
 
+    def with_text_from_prompt(self, prompt: Prompt, target: str) -> "Explanation":
+        items: List[Union[
+            TextPromptItemExplanation,
+            ImagePromptItemExplanation,
+            TargetPromptItemExplanation,
+            TokenPromptItemExplanation,
+        ]] = []
+        for item_index, item in enumerate(self.items): 
+            if isinstance(item, TextPromptItemExplanation):
+                # separate variable to fix linting error
+                prompt_item = prompt.items[item_index]
+                if isinstance(prompt_item, Text): 
+                    items.append(item.with_text(prompt_item))
+                else:
+                    items.append(item)
+            elif isinstance(item, TargetPromptItemExplanation):
+                items.append(item.with_text(target))
+            else:
+                items.append(item)
+        return Explanation(
+            target=self.target,
+            items=items,
+        )
+
+
 
 class ExplanationResponse(NamedTuple):
     model_version: str
@@ -372,6 +441,15 @@ class ExplanationResponse(NamedTuple):
     ) -> "ExplanationResponse":
         mapped_explanations = [
             explanation.with_image_prompt_items_in_pixels(prompt)
+            for explanation in self.explanations
+        ]
+        return ExplanationResponse(self.model_version, mapped_explanations)
+    
+    def with_text_from_prompt(
+        self, request: ExplanationRequest
+    ) -> "ExplanationResponse":
+        mapped_explanations = [
+            explanation.with_text_from_prompt(request.prompt, request.target)
             for explanation in self.explanations
         ]
         return ExplanationResponse(self.model_version, mapped_explanations)