+++ tests/test_explanation.py
@@ -1,4 +1,6 @@
 from pathlib import Path
+from aleph_alpha_client.explanation import TargetScoreWithRaw, TextScoreWithRaw
+from aleph_alpha_client.prompt import Tokens
 import pytest
 from aleph_alpha_client import (
     ControlTokenOverlap,
@@ -17,9 +19,7 @@ from aleph_alpha_client import (
 
 from tests.common import (
     sync_client,
-    client,
     model_name,
-    model,
     async_client,
 )
 
@@ -37,7 +37,7 @@ async def test_can_explain_with_async_client(
                     "I am a programmer###I am French###I don't like pizza###My favourite food is"
                 ),
                 # " My favorite food is"
-                [4014, 36316, 5681, 387],
+                Tokens.from_token_ids([4014, 36316, 5681, 387]),
             ]
         ),
         target=" pizza with cheese",
@@ -65,7 +65,7 @@ def test_explanation(sync_client: Client, model_name: str):
                 img,
                 Text.from_text("I am a programmer and French. My favourite food is"),
                 # " My favorite food is"
-                [4014, 36316, 5681, 387],
+                Tokens.from_token_ids([4014, 36316, 5681, 387]),
             ]
         ),
         target=" pizza with cheese",
@@ -100,7 +100,7 @@ def test_explanation_auto_granularity(sync_client: Client, model_name: str):
                 img,
                 Text.from_text("I am a programmer and French. My favourite food is"),
                 # " My favorite food is"
-                [4014, 36316, 5681, 387],
+                Tokens.from_token_ids([4014, 36316, 5681, 387]),
             ]
         ),
         target=" pizza with cheese",
@@ -123,7 +123,7 @@ def test_explanation_of_image_in_pixels(sync_client: Client, model_name: str):
                 img,
                 Text.from_text("I am a programmer and French. My favourite food is"),
                 # " My favorite food is"
-                [4014, 36316, 5681, 387],
+                Tokens.from_token_ids([4014, 36316, 5681, 387]),
             ]
         ),
         target=" pizza with cheese",
@@ -143,20 +143,16 @@ def test_explanation_of_image_in_pixels(sync_client: Client, model_name: str):
     )
 
 
-@pytest.mark.skip("not yet implemented")
 def test_explanation_of_text_in_prompt_relativ_indeces(
     sync_client: Client, model_name: str
 ):
-    image_source_path = Path(__file__).parent / "dog-and-cat-cover.jpg"
-    img = Image.from_image_source(image_source=str(image_source_path))
-
     request = ExplanationRequest(
         prompt=Prompt(
             [
-                img,
                 Text.from_text("I am a programmer and French. My favourite food is"),
+
                 # " My favorite food is"
-                [4014, 36316, 5681, 387],
+                Tokens.from_token_ids([4014, 36316, 5681, 387]),
             ]
         ),
         target=" pizza with cheese",
@@ -166,12 +162,18 @@ def test_explanation_of_text_in_prompt_relativ_indeces(
 
     explanation = sync_client._explain(request, model=model_name)
 
-    # explanation = explanation.with_text_prompt_items_absolute(request.prompt)
+    explanation = explanation.with_text_from_prompt(request)
     assert len(explanation.explanations) == 3
-    assert all([len(exp.items) == 4 for exp in explanation.explanations])
+    assert all([len(exp.items) == 3 for exp in explanation.explanations])
     assert all(
         [
-            isinstance(image_score, ImageScore) and isinstance(image_score.left, int)
-            for image_score in explanation.explanations[0].items[0].scores
+            isinstance(raw_text_score, TextScoreWithRaw) and isinstance(raw_text_score.text, str)
+            for raw_text_score in explanation.explanations[0].items[0].scores
+        ]
+    )
+    assert all(
+        [
+            isinstance(raw_text_score, TargetScoreWithRaw) and isinstance(raw_text_score.text, str)
+            for raw_text_score in explanation.explanations[1].items[2].scores
         ]
     )